Add basic clock
There is a clock rendered on screen. It has a face, it WILL have a hand,
and it has a shade which covers the portion of the face that has already
been covered by the hand.

Currently the shade is opaque-- it should be transluscent.

There are no ticks on the clock face yet. There will be eventually.

To render the mesh, the spritebatch needs to be ended and begun in order
to have the occlusion work as intended. This means each frame will have
at least two render cycles if the clock is present, and that seems
pretty bad. Hopefully this can be fixed.