Clock is prettier, has a hand
There is now a pseudo-transluscent shade that covers parts of the clock
where the hand has swept past. In other news, there is a hand on the
clock which ticks around the face.

The clock hand icon is currently twice as tall as it needs to be,
because I can't figure out how to get a sprite to rotate around a point
that is not the center of the sprite. To be fixed, if I feel like it.

There is still an issue with rendering a mesh in transparent colors, so
currently the color has just been set manually to a "slightly darker
than the face" color. If the face ever gets tick marks, this could be a
problem.